cartoon-camera/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
Pillow
torch
torchvision

